.gitlab.yml: Unpin coala version
The coala docker is currently pinned to 0.8, while 0.11
is about to be released.

Switch to using a common Python 3.5 docker image, as the
coala docker is very large in order to support bears not
used on this project.

Fix several PEP8 violations, and ignore pycodestyle rule E731
which requires converting lambdas to functions.